% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_P_Values_Connectome.r
\name{Add_P_Values_Connectome}
\alias{Add_P_Values_Connectome}
\title{This function takes the p-values calculated with 'P_Values_Connectome'
and adds them to an already-calculated connectome.  These two tasks are split up because the
unthresholded Wilcoxon rank test can take a long time}
\usage{
Add_P_Values_Connectome(connectome, p_values)
}
\arguments{
\item{p_values}{The output from 'P_Values_Connectome'}
}
\description{
This function takes the p-values calculated with 'P_Values_Connectome'
and adds them to an already-calculated connectome.  These two tasks are split up because the
unthresholded Wilcoxon rank test can take a long time
}
