% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Affinity.r
\name{Affinity}
\alias{Affinity}
\title{This function makes an affinity matrix from a connectomic file}
\usage{
Affinity(connectome, meth.use = "weight_sc", per.exp = 0,
  nodes.include)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{meth.use}{What to sum. 'num' == number of edges, 'weight' == normalized expression weight, 'weight_sc' [DEFAULT] == scaled expression weight}

\item{per.exp}{Percent expression threshold. Defaults to 0.0}

\item{nodes.include}{Character vector of nodes to perform affinity analysis over}
}
\description{
This function makes an affinity matrix from a connectomic file
}
