% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate.r
\name{Annotate}
\alias{Annotate}
\title{Annotate}
\usage{
Annotate(connectome)
}
\arguments{
\item{connectome}{A connectomic edgelist}
}
\description{
This function adds a column called 'mode' with a category for each L-R signaling interaction.
These categorizations are based on literature review and should be updated and broken down with further specificity as more information is learned.
Many of these molecules can likely be classified under multiple modes, currently this function does not support this.
}
