% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate2.r
\name{Annotate2}
\alias{Annotate2}
\title{Annotate}
\usage{
Annotate2(connectome)
}
\arguments{
\item{connectome}{A connectomic edgelist}
}
\description{
This function adds a column called 'mode' with a category for each L-R signaling interaction.
These categorizations are based on literature review and expert opinion.
Many of these molecules can likely be classified under multiple modes --  currently this function does not support this.
}
