% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCellScatter.r
\name{CellCellScatter}
\alias{CellCellScatter}
\title{CellCellScatter}
\usage{
CellCellScatter(connectome, cell.1, cell.2, lab.thresh = 1)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{cell.1}{The source cell}

\item{cell.2}{The receiving cell}

\item{lab.thresh}{Threshold for labeling of plot, applied to both x- and y- axes. Defaults to 1.}
}
\description{
This function takes a connectome and two cells and plots the edges based on ligand versus receptor weights
}
