% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircosDiff.R
\name{CircosDiff}
\alias{CircosDiff}
\title{CircosDiff}
\usage{
CircosDiff(differential.connectome, cols.use = NULL, lab.cex = 1,
  edge.color.by.source = T, ...)
}
\arguments{
\item{differential.connectome}{A differential edgelist made using DifferentialConnectome}

\item{cols.use}{Optional. Colors for plotting nodes.}

\item{lab.cex}{Text size for gene names}

\item{edge.color.by.source}{Default TRUE - edges will be colored by their source cell type. If false, edges will be colored by receiving cell instead.}
}
\description{
A Circos Plot for a differential edgelist made using DifferentialConnectome
}
