% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareCentrality.R
\name{CompareCentrality}
\alias{CompareCentrality}
\title{CompareCentrality}
\usage{
CompareCentrality(connectome.list, features = NULL,
  modes.include = NULL, nodes.include = NULL,
  weight.attribute = "weight_sc")
}
\arguments{
\item{connectome.list}{A NAMED list of connectomic edgelists. The output plot will be split by the names of the list.}

\item{features}{Genes of interest to generate the network for centrality analysis.}

\item{modes.include}{Modes to include in centrality analysis.}

\item{nodes.include}{Cell types to include in the network analysis.}
}
\description{
CompareCentrality
}
