% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connectome6.r
\name{Connectome6}
\alias{Connectome6}
\title{Connectome Version 6}
\usage{
Connectome6(object, use.supported_known = T, use.supported = F,
  use.all = F, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{use.supported_known}{Determines which set of ligand receptor pairs to use. Defaults to TRUE}

\item{use.supported}{If TRUE, overrides the above and includes all literature supported ligand receptor pairs.}

\item{use.all}{If TRUE, overrides the above and includes all possible ligand receptor pairs from the FANTOM5 database}
}
\description{
Creates a connectomic edgelist from a Seurat 2.0 object
}
