% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connectome8.r
\name{Connectome8}
\alias{Connectome8}
\title{Connectome Version 8}
\usage{
Connectome8(object, species, use.supported_known = F,
  use.supported = F, use.all = T, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{species}{The species of the object that is being processed.  Allows 'human','mouse','rat', or 'pig'}

\item{use.supported_known}{Determines which set of ligand receptor pairs to use. Defaults to TRUE}

\item{use.supported}{If TRUE, overrides the above and includes all literature supported ligand receptor pairs.}

\item{use.all}{If TRUE, overrides the above and includes all possible ligand receptor pairs from the FANTOM5 database}
}
\description{
Creates a connectomic edgelist from a Seurat 3.0 object.  Allows orphan ligand and receptors and fully non-expressed pairs. Useful for comparative connectomics.
}
