% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connectome9.r
\name{Connectome9}
\alias{Connectome9}
\title{Connectome Version 9}
\usage{
Connectome9(object, species, include.putative = T,
  include.excluded = F, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{species}{The species of the object that is being processed.  Requires input, and allows 'human','mouse','rat', or 'pig'}

\item{include.putative}{Default TRUE. Includes ligand-receptor pairs deemed putative in FANTOM5 database.}

\item{include.all}{Default FALSE. If TRUE, includes gene pairs labeled EXCLUDED in FANTOM5 database.  See ncomms8866 .rda file for qualifications for exclusion.}
}
\description{
Creates a connectomic edgelist from a Seurat 3.0 object.  Allows orphan ligand and receptors and fully non-expressed pairs. Useful for comparative connectomics.
}
