% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectomeFold.R
\name{ConnectomeFold}
\alias{ConnectomeFold}
\title{Connectome Fold Change Comparison}
\usage{
ConnectomeFold(connect.1, connect.2)
}
\arguments{
\item{connect.1}{A connectome from a system}

\item{connect.2}{A connectome from a different system, to be compared to connect.1}
}
\description{
Creates a fold-change connectome from two input connectomes, inlcuding MAX outputs for things present in only one
}
