% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conserved_Edges_WR.R
\name{Conserved_Edges_WR}
\alias{Conserved_Edges_WR}
\title{Conserved_Edges}
\usage{
Conserved_Edges_WR(con.list)
}
\arguments{
\item{con.list}{A list of connectomes to calculate conservation across}
}
\description{
This function takes multiple connectomic edgelists as input and returns a single master connectome identifying conserved edges and average properties
}
