% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_Markers_to_Homologues.r
\name{Convert_Markers_to_Homologues}
\alias{Convert_Markers_to_Homologues}
\title{Convert_to_Homologues}
\usage{
Convert_Markers_to_Homologues(mark, old.species, new.species = "human")
}
\arguments{
\item{mark}{A marker list (the output from FindMarkers or FindAllMarkers in Seurat)}

\item{old.species}{The original species of the object (can be 'mouse','rat','pig',or 'human')}

\item{new.species}{The desired species mapping. Defaults to 'human'}
}
\description{
This function converts a MARKER LIST to a homologue object.  Can be used to convert in any direction between mouse, rat, pig, and human.
}
