% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_to_Homologues_v2.r
\name{Convert_to_Homologues_v2}
\alias{Convert_to_Homologues_v2}
\title{Convert_to_Homologues_v2}
\usage{
Convert_to_Homologues_v2(object, old.species, new.species = "human")
}
\arguments{
\item{object}{A Seurat 3.0 object. This code will not work with a Seurat 2.0 object as the data slot organization is different}

\item{old.species}{The original species of the object (can be 'mouse','rat','pig',or 'human')}

\item{new.species}{The desired species mapping. Defaults to 'human'}
}
\description{
This function converts a SEURAT OBJECT to a homologue object.  Can be used to convert in any direction between mouse, rat, pig, and human.
}
