% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseModalNetwork.R
\name{DenseModalNetwork}
\alias{DenseModalNetwork}
\title{DenseModalNetwork}
\usage{
DenseModalNetwork(connectome, MOI)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{MOI}{The signaling mode of interest to include in the network analysis and subsequent plotting}
}
\description{
This function takes a connectomic edgelist and creates a visualization of the cell-cell network for a given mode with edges that are the sum of all weights. Must have modal annotations for this to work.
}
