% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffEdgeDotPlot.R
\name{DiffEdgeDotPlot}
\alias{DiffEdgeDotPlot}
\title{DiffEdgeDotPlot}
\usage{
DiffEdgeDotPlot(
  differential.connectome,
  features = NULL,
  sources.include = NULL,
  targets.include = NULL,
  min.score = NULL,
  min.pct = NULL,
  verbose = T,
  infinity.to.max = T,
  ...
)
}
\arguments{
\item{differential.connectome}{A differential connectome object}

\item{features}{Gene of interest. Output will be limited to edges including these specific genes.}

\item{sources.include}{Source nodes of interest. Output will be limited to edges coming from these sources.}

\item{targets.include}{Target nodes of interest. Output will be limited to edges landing on these targets.}

\item{min.score}{Default NULL. Will limit output to edges with a differential score greater than this value.}

\item{min.pct}{Default NULL. Threshold to return clusterwise observations for both ligand and receptor. Only needs to be satisfied in connect.1 OR in connect.2.}

\item{verbose}{Whether to output feedback to user}

\item{infinity.to.max}{Default TRUE.  If TRUE, will create a pseudo value to replace values of "Inf"}
}
\description{
Plotting function to make a DotPlot of differential edges for a network comparison. This function first finds all edges meeting the desired thresholding criteria, and then plots complete information regarding all mechanisms and celltype vectors implicated.
}
