% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialConnectome.R
\name{DifferentialConnectome}
\alias{DifferentialConnectome}
\title{DifferentialConnectome}
\usage{
DifferentialConnectome(connect.1, connect.2)
}
\arguments{
\item{connect.1}{A connectome from a system}

\item{connect.2}{A connectome from a different system, to be compared to connect.1}
}
\description{
Creates a fold-change connectome from two input connectomes.  Must be node-aligned, from the same reference mapping, and unfiltered. ('edge' columns must contain identical entries, though not necessarily in the same order.)
}
