% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialScoringPlot.R
\name{DifferentialScoringPlot}
\alias{DifferentialScoringPlot}
\title{DifferentialScoringPlot}
\usage{
DifferentialScoringPlot(differential.connectome, min.score = NULL)
}
\arguments{
\item{differential.connectome}{A differential connectome, made with DifferentialConnectome. May be filtered as desired prior to plotting.}

\item{min.score}{Default NULL. Threshold to prioritize only strongly perturbed edges.}
}
\description{
Currently in beta testing. Creates x3 aligned heatmaps allowing visualization of ligand, receptor, and perturbation scores for a given cell-system of interest.
}
