% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOR.R
\name{DOR}
\alias{DOR}
\title{DOR}
\usage{
DOR(object, ident, features)
}
\arguments{
\item{object}{A connectomic object}

\item{ident}{Cell identity of interest}

\item{features}{Features of interest to calculate DOR}
}
\description{
DOR
}
