% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeDotPlot.R
\name{EdgeDotPlot}
\alias{EdgeDotPlot}
\title{EdgeDotPlot}
\usage{
EdgeDotPlot(connectome, ...)
}
\arguments{
\item{connectome}{A connectomic object}

\item{...}{Arguments passed to FilterConnectome}
}
\description{
Plotting function to make a DotPlot of edges within a specified network. This function first finds all edges meeting the desired thresholding criteria, and then plots complete information regarding all mechanisms and celltype vectors implicated.
}
