% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveOrganization.r
\name{HiveOrganization}
\alias{HiveOrganization}
\title{HiveOrganization}
\usage{
HiveOrganization(connectome)
}
\arguments{
\item{connectome}{A connectomic edgelist}
}
\description{
This function processes connectome data to act as input for http://wodaklab.org/hivegraph/graph, based on 'mode' categories. Will write a set of .txt files directly to working directory.
}
