% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalNetwork.r
\name{ModalNetwork}
\alias{ModalNetwork}
\title{ModalNetwork}
\usage{
ModalNetwork(connectome, MOI)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{MOI}{The signaling mode of interest to include in the network analysis and subsequent plotting}
}
\description{
This function takes a connectomic edgelist and creates a visualization of the cell-cell network for a given mode. Must have modal annotations for this to work.
}
