% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModeScatter.r
\name{ModeScatter}
\alias{ModeScatter}
\title{ModeScatter}
\usage{
ModeScatter(connectome, MOI = "Wnt", weight = "weight_sc",
  rec.wt = "recept.scale", lig.wt = "ligand.scale")
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{MOI}{The mode of interest (i.e., 'Wnt')}

\item{weight}{The column name for the edgeweight of interest (determines dot sizes). Defaults to weight_sc}

\item{rec.wt}{The column name for receptor expression (x axis). Defaults to recept.scale}

\item{lig.wt}{The column name for ligand expression (y axis). Defaults to ligand.scale}

\item{cell2}{The receiving cell}
}
\description{
This function takes a connectome and plots the interactions within a single signaling modality. x-axis is receptor expression and y-axis is ligand expression. Color is exact mechanism, label is cell-cell vector.
}
