% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NearestNeighbor.r
\name{NearestNeighbor}
\alias{NearestNeighbor}
\title{NearestNeighbor}
\usage{
NearestNeighbor(connectome, NOI, return = F)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{NOI}{The nodes of interest to include}
}
\description{
This function takes a connectome and plots a visualization of cell-cell connectivity showing the sum of the weights of all edges between each cell combo.
}
