% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NearestNeighborModal.r
\name{NearestNeighborModal}
\alias{NearestNeighborModal}
\title{Nearest Neighbor}
\usage{
NearestNeighborModal(connectome.list, NOI, mode_plot = "Growth factors")
}
\arguments{
\item{connectome.list}{A list of connectomic edgelists}

\item{NOI}{The nodes of interest to include in the network analysis and subsequent plotting}

\item{mode_plot}{The mode of interest to plot}
}
\description{
This function makes a mode-specific dotplot showing nearest neighbors within a networks
Can compare a LIST of connectomes
requires weight column to be called 'weight_sc' <-- should be revised to be more flexible
}
