% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworksCompare.r
\name{NetworksCompare}
\alias{NetworksCompare}
\title{NetworksCompare}
\usage{
NetworksCompare(aff.list, output = c("full"), ...)
}
\arguments{
\item{aff.list}{A list of affinity matrices made with Affinity}
}
\description{
This function takes a list of affinity matrices (made with 'Affinity') and plots the networks side-by-side.
Output will be a PDF in your working directory.
}
