% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworksCompare2.r
\name{NetworksCompare2}
\alias{NetworksCompare2}
\title{NetworksCompare2}
\usage{
NetworksCompare2(con.list, MOI, weight.attribute = "weight_sc", ...)
}
\arguments{
\item{con.list}{A list of connectomes}

\item{MOI}{The mode of interest to plot}
}
\description{
This function takes a list of connectomes and plots the networks side by side. Performs on a mode of interest (MOI)
Output will be an image
}
