% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheDotPlot.r
\name{NicheDotPlot}
\alias{NicheDotPlot}
\title{NicheDotPlot}
\usage{
NicheDotPlot(connectome, NOI, return = F)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{NOI}{A single node of interest. This function will plot an illustration of the niche of this node.}
}
\description{
This function takes a connectomic edgelist and creates a dot plot of a given niche, organized by source cell and mode. The y-axis is binned by source cell type,
and the x-axis is the sum of the weight of all edges for a given ligand made by each cell.
}
