% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantify_Conversion.r
\name{Quantify_Conversion}
\alias{Quantify_Conversion}
\title{Quantify_Conversion}
\usage{
Quantify_Conversion(gene.list, old.species, new.species = "human")
}
\arguments{
\item{gene.list}{A gene list}

\item{old.species}{The original species of the object (can be 'mouse','rat','pig',or 'human')}

\item{new.species}{The desired species mapping. Defaults to 'human'}
}
\description{
This function quantifies the homologue conversion of a GENE LIST to a homologue object.
}
