% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalScatter.r
\name{SignalScatter}
\alias{SignalScatter}
\title{SignalScatter}
\usage{
SignalScatter(connectome, features)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{features}{Gene of interest. All edges containing these features will be plotted.}
}
\description{
Scatter plot of signaling vectors involving features of interest. Returns a ggplot object, colored by mechanism, labeled by cell-cell vector.
}
