% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threshold_Master.R
\name{Threshold_Master}
\alias{Threshold_Master}
\title{Threshold_Master}
\usage{
Threshold_Master(connectome, pct.thresh = 0.05, p.thresh = 0.05,
  scale.thresh = NULL)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{pct.thresh}{Percent of cluster expressing marker}

\item{p.thresh}{P-value cutoff, defaults to 0.05}

\item{scale.thresh}{Scale value cutoff, defaults to NULL}
}
\description{
This function thresholds a connectome to only include those edges which meet the input parameters
}
