% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCellScatter.r
\name{CellCellScatter}
\alias{CellCellScatter}
\title{CellCellScatter}
\usage{
CellCellScatter(connectome, cell1, cell2, weight = "weight_sc",
  rec.wt = "recept.scale", lig.wt = "ligand.scale")
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{cell1}{The source cell}

\item{cell2}{The receiving cell}

\item{weight}{The column name for the edgeweight of interest (determines dot sizes). Defaults to 'weight_sc'}

\item{rec.wt}{The column name for receptor expression (x axis). Defaults to 'recept.scale'}

\item{lig.wt}{The column name for ligand expression (y axis). Defaults to 'ligand.scale'}
}
\description{
This function takes a connectome and two cells and plots the interactions based on ligand versus receptor weights
}
