% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircosDiff.R
\name{CircosDiff}
\alias{CircosDiff}
\title{CircosDiff}
\usage{
CircosDiff(
  differential.connectome,
  features = NULL,
  sources.include = NULL,
  targets.include = NULL,
  min.score = NULL,
  min.pct = NULL,
  verbose = T,
  infinity.to.max = T,
  edge.color.by.source = T,
  cols.use = NULL,
  lab.cex = 1,
  title = NULL,
  small.gap = 1,
  big.gap = 10
)
}
\arguments{
\item{differential.connectome}{A differential connectome, made with DifferentialConnectome. May be filtered as desired prior to plotting.}

\item{features}{Gene of interest. Output will be limited to edges including these specific genes.}

\item{sources.include}{Source nodes of interest. Output will be limited to edges coming from these sources.}

\item{targets.include}{Target nodes of interest. Output will be limited to edges landing on these targets.}

\item{min.score}{Default NULL. Will limit output to edges with a differential score greater than this value.}

\item{min.pct}{Default NULL. Threshold to return clusterwise observations for both ligand and receptor. Only needs to be satisfied in connect.1 OR in connect.2.}

\item{verbose}{Whether to output feedback to user}

\item{infinity.to.max}{Default TRUE.  If TRUE, will create a pseudo value to replace values of "Inf"}

\item{edge.color.by.source}{Default TRUE - edges will be colored by their source cell type. If false, edges will be colored by receiving cell instead.}

\item{cols.use}{Optional. Colors for plotting nodes.}

\item{lab.cex}{Text size for gene names}

\item{title}{Character string for title of plot.}

\item{small.gap}{Default 1. Amount of distance between sectors.  If the number of edges is very large, this will have to be reduced in size.}

\item{big.gap}{Default 10. Amount of distance between the source cells and the target cells (top and bottom arc of graph).  If the number of edges is very large, this can be reduced in size in addition to 'small.gap'}
}
\description{
Currently in beta testing. A Circos Plot for a differential edgelist made using DifferentialConnectome. Note that negative log fold changes cannot be plotted; therefore, this plot uses the 'score' for the differential comparison, which is always positive in proportion to perturbation.
}
