% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareCentrality.R
\name{CompareCentrality}
\alias{CompareCentrality}
\title{CompareCentrality}
\usage{
CompareCentrality(connectome.list, weight.attribute = "weight_norm",
  min.z = NULL, ...)
}
\arguments{
\item{connectome.list}{A named list of connectomic edgelists. The output plot will be split by the names of the list.}

\item{weight.attribute}{Column to use to define edgeweights for network analysis.  'weight_sc' or 'weight_norm'. Defaults to 'weight_norm'}

\item{min.z}{Minimum z-score for ligand and receptor.}

\item{...}{Arguments passed to FilterConnectome.  Will be applied to each connectome within the list.}
}
\description{
Takes a list of connectomes and will compare sending- and receiving- centrality, side-by-side, for the given networks, or requested network subsets.
}
