% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialScoringPlot.R
\name{DifferentialScoringPlot}
\alias{DifferentialScoringPlot}
\title{DifferentialScoringPlot}
\usage{
DifferentialScoringPlot(differential.connectome, features = NULL,
  sources.include = NULL, targets.include = NULL, min.score = NULL)
}
\arguments{
\item{differential.connectome}{A differential connectome, made with DifferentialConnectome. May be filtered as desired prior to plotting.}

\item{features}{Gene of interest. Output will be limited to edges including these specific genes.}

\item{sources.include}{Source nodes of interest. Output will be limited to edges coming from these sources.}

\item{targets.include}{Target nodes of interest. Output will be limited to edges landing on these targets.}

\item{min.score}{Default NULL. Will limit output to edges with a differential score greater than this value.}
}
\description{
Currently in beta testing. Creates x3 aligned heatmaps allowing visualization of ligand, receptor, and perturbation scores for a given cell-system of interest.
}
