% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialScoringPlot.R
\name{DifferentialScoringPlot}
\alias{DifferentialScoringPlot}
\title{DifferentialScoringPlot}
\usage{
DifferentialScoringPlot(
  differential.connectome,
  features = NULL,
  sources.include = NULL,
  targets.include = NULL,
  min.score = NULL,
  min.pct = NULL,
  verbose = T,
  infinity.to.max = T,
  aligned = F
)
}
\arguments{
\item{differential.connectome}{A differential connectome, made with DifferentialConnectome. May be filtered as desired prior to plotting.}

\item{features}{Gene of interest. Output will be limited to edges including these specific genes.}

\item{sources.include}{Source nodes of interest. Output will be limited to edges coming from these sources.}

\item{targets.include}{Target nodes of interest. Output will be limited to edges landing on these targets.}

\item{min.score}{Default NULL. Will limit output to edges with a differential score greater than this value.}

\item{min.pct}{Default NULL. Threshold to return clusterwise observations for both ligand and receptor. Only needs to be satisfied in connect.1 OR in connect.2.}

\item{verbose}{Whether to output feedback to user}

\item{infinity.to.max}{Default TRUE.  If TRUE, will create a pseudo value to replace values of "Inf"}

\item{aligned}{Default FALSE. If TRUE, will create edge-aligned heatmaps (duplicate rows and columns in first two plots, to dimension map all three plots)}
}
\description{
Currently in beta testing. Creates x3 aligned heatmaps allowing visualization of ligand, receptor, and perturbation scores for a given cell-system of interest.
}
