% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterConnectome.R
\name{FilterConnectome}
\alias{FilterConnectome}
\title{FilterConnectome}
\usage{
FilterConnectome(connectome, min.pct = 0.1, min.exp = 0, min.z = 0,
  max.p = 0.01, ...)
}
\arguments{
\item{connectome}{A connectomic edgelist output from CreateConnectome}

\item{min.pct}{Minimum fraction of cells within a given cluster expressing the ligand or receptor. Defaults to 0.10, allows NULL.}

\item{min.exp}{Minimum normalized expression level of ligand and receptor. Defaults to 0, allows NULL.}

\item{min.z}{Minimum z-score for ligand and receptor. Defaults to 0, allows NULL.}

\item{max.p}{Maximum p-value for ligand and receptor. Defaults to 0.01. Requires prior p-value calculation.}
}
\description{
Filters a connectomic edgelist output from CreateConnectome according to user inputs. Defaults are set to reasonable initial parameters for data clean-up and exploration.
}
