% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalDotPlot.r
\name{ModalDotPlot}
\alias{ModalDotPlot}
\title{ModalDotPlot}
\usage{
ModalDotPlot(connectome, NOI, return = F)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{NOI}{The nodes of interest to include in the network analysis and subsequent plotting}
}
\description{
This function takes a connectomic edgelist and creates a source and a sink mode- and cell- organized dot plot. The y-axis is binned by mode,
and the x-axis is the sum of the weight of all edges for each mode made by each cell. Points are binned by cell type, with the size of the point
correlating to the Kleinberg hub score (for source graph) and Kleinberg authority score (for sink graph)
}
