% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalDotPlot.r
\name{ModalDotPlot}
\alias{ModalDotPlot}
\title{ModalDotPlot}
\usage{
ModalDotPlot(connectome, nodes = NULL)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{nodes}{The nodes (cell identities) of interest to include in the network analysis and subsequent plotting. Defaults to all nodes.}
}
\description{
This function takes a connectomic edgelist and creates a source and a sink mode- and cell- organized dot plot. The y-axis is the discrete variable 'mode',
and the x-axis is the sum of the weight (weight_sc column) of all edges for each mode made by each cell. Points are organized by cell type, with the size of the point
correlating to the Kleinberg hub score (for source graph) and Kleinberg authority score (for sink graph)
}
