% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlot.r
\name{NetworkPlot}
\alias{NetworkPlot}
\title{NetworkPlot}
\usage{
NetworkPlot(connectome, weight.attribute = "weight_sc", title = NULL,
  cols.use = NULL, include.all.nodes = F, ...)
}
\arguments{
\item{connectome}{A connectomic object, ideally filtered to only edges of interest.}

\item{weight.attribute}{Column to use to define edgeweights for network analysis. 'weight_sc' or 'weight_norm'. Defaults to 'weight_sc'}

\item{title}{Description of the network being plotted}

\item{cols.use}{Optional. Colors for plotting nodes.}

\item{...}{Arguments passed to FilterConnectome}
}
\description{
Creates a network plot of a connectomic object.  Wrapper for igraph functionality.
}
