% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlot.r
\name{NetworkPlot}
\alias{NetworkPlot}
\title{NetworkPlot}
\usage{
NetworkPlot(connectome, features = NULL,
  weight.attribute = "weight_sc", title = NULL, modes.include = NULL,
  ...)
}
\arguments{
\item{connectome}{A connectomic object, ideally filtered to only edges of interest.}

\item{features}{Genes of interest. All edges containing these features will be plotted.}

\item{weight.attribute}{The desired column to use for edgeweights. Defaults to 'weight_sc'}

\item{title}{Description of the network being plotted}

\item{modes.include}{Modes to be plotted. Defaults to all modes. Can be used to look at a narrow category of signaling.}
}
\description{
Creates a network plot of a connectomic object.  Wrapper for igraph functionality.
}
