% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalScatter.r
\name{SignalScatter}
\alias{SignalScatter}
\title{SignalScatter}
\usage{
SignalScatter(connectome, feature)
}
\arguments{
\item{connectome}{A connectomic edgelist}

\item{feature}{Gene of interest. All edges containing this feature will be plotted.}
}
\description{
Scatter plot of signaling vectors involving features of interest.
}
